Fixed a chordline in See That My Grave Is Kept Clean -- Eyolf, please check!
We had as the last line of the second verse
  D         c  /b       D
  Waiting on my burying ground.
I changed that to
  D       /c    /b      D
  Waiting on my burying ground.
(my regex doesn't match this as a chordline without a slash before the c :))
I listened to the record, but my ears can't tell what he really plays when.
Could as well be some interpretation of
  D     /c /b           D
  Waiting on my burying ground.
meaning the bass line changes between the words, or whatever.